backend/pyproject.toml
[project]
name = "fastapi-demo-app-backend"
version = "0.1.0"
description = "Python FastAPI server that integrates Wristband authentication."
authors = [
    {name = "Wristband", email = "support@wristband.dev"},
]
license = {text = "MIT"}
requires-python = ">=3.9.2,<4.0"
dependencies = [
    "fastapi>=0.115.12,<0.116.0",
    "httpx>=0.28.1,<0.29.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "uvicorn>=0.34.0,<0.35.0",
    "wristband-fastapi-auth (>=1.0.0,<2.0.0)",
    "firebase-admin (>=6.9.0,<7.0.0)",
    "pytest>=7.0.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-mock>=3.10.0,<4.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "cryptography>=44.0.3,<46.0.0",
]

[tool.poetry]
packages = [
    {include = "src"},
    {include = "auth", from = "src"},
    {include = "clients", from = "src"},
    {include = "database", from = "src"},
    {include = "middleware", from = "src"},
    {include = "models", from = "src"},
    {include = "routes", from = "src"},
    {include = "utils", from = "src"},
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (requires live API access)",
    "unit: marks tests as unit tests (mocked, no external dependencies)",
    "slow: marks tests as slow running",
]
addopts = "-v --tb=short --strict-markers"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
